// src/api/user/user.types.d.ts
export type LeadCataloguePackage = {
    id: number;
    lead_id: number;
    package_id: number;
    package_start_date: string;
    package_end_date: string;
    is_active: number;
    created_at: string;
    updated_at: string;
    is_aia: string;
    is_direct_assigned: boolean;
    catalogue_package: unknown;
    lead_catalogue_package_feature: LeadCataloguePackageFeature[];
};
export declare enum FeatureCategoryId {
    NoOfCalls = 1,
    RayaFlatDiscount = 2,
    LabDiscount = 3,
    MBEPrice = 4,
    Delivery = 5,
    MedicineDiscount = 6,
    NoOfInsuranceOrder = 7,
    NoOfCallsTalkLater = 8,
    PartnerAndCouponManagement = 9,
    NoOfTalkGlobalCalls = 10,
    NoOfTalkLaterCalls = 11,
    Group = 12,
    Claim = 13,
    TalkLaterDoctorWise = 14
}
export type LeadCataloguePackageFeature = {
    id: number;
    lead_catalogue_package_id: number;
    catalogue_package_feature_id: number;
    category_id: FeatureCategoryId;
    feature_title: string;
    feature_value: string | null;
    is_active: number;
    created_at: string;
    updated_at: string;
    original_feature_value: string | null;
    insurance_id: number | null;
    data: any | null;
    is_aia: boolean | null;
    is_flash_credit_usable: boolean | null;
};
export type ClaimFeatures = {
    value: string;
    medicine: boolean;
    lab: boolean;
    raya: boolean;
    talkNow: boolean;
    talkLater: boolean;
    talkGlobal: boolean;
    talkFamily: boolean;
    talkFamilyMother: boolean;
    talkFamilyFather: boolean;
    talkFamilyBrother: boolean;
    talkFamilySister: boolean;
    talkFamilySpouse: boolean;
    talkFamilyChild: boolean;
    creditType: number;
};